Silence 30 paper_trail deprecations during CI
Example warning was:
```
DEPRECATION WARNING: PaperTrail.track_associations has not been set. As of PaperTrail 5, it defaults to false. Tracking associations is an
experimental feature so we recommend setting PaperTrail.config.track_associations = false in your config/initializers/paper_trail.rb
```

Note: not using force for config, because if a user has generated an
app and modified it, we want them to get the conflict/diff prompts, not
just lose their config.